+++ mls_chapter_images/2d_cpw_heterogeneous_sweep_marmousi/adapted_mesh_marmousi.png
