+++ mls_chapter_images/2d_cpw_heterogeneous_sweep_marmousi/mls_cpw_results_heterogeneous_mltriangles.png
