+++ mls_chapter_images/2d_cpw_heterogeneous_sweep_marmousi/organize_data.py
