+++ mls_chapter_images/2d_cpw_heterogeneous_sweep_marmousi/p2_cpw_results_fine.txt
