+++ mls_chapter_images/2d_cpw_heterogeneous_sweep_marmousi/p4_cpw_results_refined_again.txt
