+++ mls_chapter_images/2d_cpw_heterogeneous_sweep_marmousi/plot_mls_2d_heterogeneous_image.py
